requirements.txt
ConfigArgParse==1.5.3
cupy_cuda111==11.6.0
imageio==2.19.3
matplotlib==3.5.3
numpy==1.21.4
opencv_python==4.5.4.60
Pillow==9.2.0
pytorch_msssim==0.2.1
scikit_image==0.17.2
torch==1.10.0+cu111
torchvision==0.11.0+cu111
tqdm==4.62.3
